+++ w/g_s/css/game-elements.css
@@ -1,0 +1,81 @@
+#gun {
+    position: absolute;
+    bottom: 20px;
+    left: 50%;
+    width: 30px;
+    height: 30px;
+    background: #444;
+    transform: translateX(-50%);
+    cursor: crosshair;
+    border-radius: 50%;
+    transform-origin: center center;
+    box-shadow: 0 0 15px rgba(0, 0, 0, 0.3);
+}
+
+.gun-barrel {
+    position: absolute;
+    top: -12px;
+    width: 8px;
+    height: 15px;
+    background: #555;
+    border-radius: 3px;
+    transform-origin: bottom center;
+    transition: transform 0.2s ease;
+    will-change: transform;
+    box-shadow: 0 0 5px rgba(0, 0, 0, 0.5);
+}
+
+.balloon {
+    --base-size: var(--base-balloon-size);
+    position: absolute;
+    width: calc(var(--base-size) * var(--size-multiplier, 2));
+    height: calc(var(--base-size) * var(--size-multiplier, 2));
+    background: red;
+    border-radius: 50%;
+    cursor: pointer;
+    box-shadow: 0 0 10px rgba(255, 255, 255, 0.1);
+}
+
+.bullet {
+    position: absolute;
+    width: 6px;
+    height: 6px;
+    background: #ffd700;
+    border-radius: 50%;
+    opacity: 0.8;
+    box-shadow: 0 0 8px #ffa500,
+                0 0 16px rgba(255, 165, 0, 0.6);
+    filter: drop-shadow(0 0 2px rgba(255, 215, 0, 0.4));
+}
+
+.particle {
+    position: absolute;
+    width: 8px;
+    height: 8px;
+    border-radius: 50%;
+    pointer-events: none;
+    box-shadow: 0 0 5px rgba(255, 255, 255, 0.3);
+    filter: brightness(1.2);
+}
+
+.pause-overlay {
+    position: fixed;
+    inset: 0;
+    background: rgba(0, 0, 0, 0.7);
+    display: none;
+    justify-content: center;
+    align-items: center;
+    z-index: 1000;
+    backdrop-filter: blur(3px);
+    -webkit-backdrop-filter: blur(3px);
+}
+
+.pause-overlay.active {
+    display: flex;
+}
+
+.pause-text {
+    color: var(--color-text);
+    font-size: 48px;
+    text-shadow: 0 0 10px rgba(255, 255, 255, 0.5);
+}
