+++ w/g_s/js/index.js
@@ -1,0 +1,6 @@
+import { Game } from './Game.js';
+
+window.addEventListener('DOMContentLoaded', () => {
+    const game = new Game();
+    window.addEventListener('unload', () => game.cleanup());
+});
